[XLA:GPU] Move the s4 unpacking sequence from llvm pass to int4->int8 pass
The patch adds the rewrite where we replace extSi + SiToFP with inline assembly right in int4 rewrite pass. As a result of that we can patch triton with pr#24 and pr#25 that speed up the memory transfer. As a result of that we get the speed improvement of int4 x bf16 gemms up to 1.44 times for some shapes.

PiperOrigin-RevId: 783708907