[xla] Change Shape::IsArray and Shape::IsArrayOrBuffer.
Currently, Shape::IsArray returns true when the shape is an array primitive and Shape::IsArrayOrBuffer returns true when the shape is an array primitive or a buffer primitive.

We rename these two interfaces to Shape::IsArrayExcludingBuffer and Shape::IsArray respectively. This can avoid the need to change many existing uses of Shape::IsArray to Shape::IsArrayOrBuffer to support the new buffer type.

PiperOrigin-RevId: 785976841